Some ingration test fixes, adds a shared_lambda_secrets_access resource for rds on pr env

+++ src/test/integration/initPayment.test.ts
@@ -7,11 +7,13 @@ describe("initialize a payment", () => {
   const transactionReferenceId = crypto.randomUUID();
 
   afterAll(async () => {
-    if (isLocal) {
+    if (isLocal && knex) {
       await TransactionModel.query()
         .where({ transactionReferenceId })
         .delete();
-      await knex.destroy();
+      if (typeof knex.destroy === 'function') {
+        await knex.destroy();
+      }
     }
   });
